Remove redundant type casts

+++ src/instruction.rs
@@ -372,7 +372,7 @@ impl Instruction {
             | Some(MemoryIndexedAndRegister(_, idx, _))
             | Some(MemoryIndexedBitAndRegister(_, idx, _, _)) => bytes.push(idx as u8),
             Some(ProgramCounterRelative(offset)) => bytes.push(offset as u8),
-            Some(PortDirect(port)) => bytes.push(port as u8),
+            Some(PortDirect(port)) => bytes.push(port),
             _ => (),
         };
 
@@ -385,7 +385,7 @@ impl Instruction {
             | Some(MemoryIndexedAndRegister(_, idx, _))
             | Some(MemoryIndexedBitAndRegister(_, idx, _, _)) => bytes.push(idx as u8),
             Some(ProgramCounterRelative(offset)) => bytes.push(offset as u8),
-            Some(PortDirect(port)) => bytes.push(port as u8),
+            Some(PortDirect(port)) => bytes.push(port),
             _ => (),
         };
